% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plink_bim_col_names}
\alias{plink_bim_col_names}
\title{col names for plink to standardize since I have been copy-pasting}
\format{
An object of class \code{character} of length 6.
}
\usage{
plink_bim_col_names
}
\description{
col names for plink to standardize since I have been copy-pasting
}
\keyword{datasets}
