% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_tools.R
\name{select_score_allele}
\alias{select_score_allele}
\title{Select score allele from a data frame}
\usage{
select_score_allele(variantTbl_proxy_selected)
}
\description{
Select score allele from a data frame
}
