% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{write_chr_tbl}
\alias{write_chr_tbl}
\title{write a table contains a chromosome column (assume unique) and use the chr as a part of file name}
\usage{
write_chr_tbl(chr_tbl, prefix, suffix, delim, col_names)
}
\description{
write a table contains a chromosome column (assume unique) and use the chr as a part of file name
}
